requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
bcrypt==4.2.1
python-multipart==0.0.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
requests>=2.31.0
python-docx==1.1.2
jinja2==3.1.4
